Needed to change profile for the check to even be found
