+++ tests/profiles/googlefonts_test.py
@@ -26,6 +26,7 @@ from fontbakery.constants import (NameID,
                                   MacintoshLanguageID,
                                   OFL_BODY_TEXT)
 from fontbakery.profiles import googlefonts as googlefonts_profile
+from fontbakery.profiles import opentype as opentype_profile
 
 check_statuses = (ERROR, FAIL, SKIP, PASS, WARN, INFO, DEBUG)
 
@@ -2276,7 +2277,7 @@ def NOT_IMPLEMENTED_test_check_fsselection():
 
 def test_check_italic_angle():
     """ Checking post.italicAngle value. """
-    check = CheckTester(googlefonts_profile,
+    check = CheckTester(opentype_profile,
                         "com.google.fonts/check/italic_angle")
 
     ttFont = TTFont(TEST_FILE("cabin/Cabin-Regular.ttf"))
@@ -2612,23 +2613,6 @@ def test_condition_familyname_with_spaces():
     from fontbakery.profiles.googlefonts_conditions import familyname_with_spaces
     assert familyname_with_spaces("OverpassMono") == "Overpass Mono"
     assert familyname_with_spaces("BodoniModa11") == "Bodoni Moda 11"
-
-
-def test_style_condition():
-    from fontbakery.profiles.shared_conditions import style
-    # VFs
-    assert style(TEST_FILE("shantell/ShantellSans[BNCE,INFM,SPAC,wght].ttf")) == "Regular"
-    assert style(TEST_FILE("shantell/ShantellSans-Italic[BNCE,INFM,SPAC,wght].ttf")) == "Italic"
-    assert style(TEST_FILE("shantell/ShantellSans-Bold[BNCE,INFM,SPAC,wght].ttf")) == "Bold"
-    assert style(TEST_FILE("shantell/ShantellSans-BoldItalic[BNCE,INFM,SPAC,wght].ttf")) == "BoldItalic"
-    # Statics
-    assert style(TEST_FILE("bad_fonts/style_linking_issues/NotoSans-Regular.ttf")) == "Regular"
-    assert style(TEST_FILE("bad_fonts/style_linking_issues/NotoSans-Italic.ttf")) == "Italic"
-    assert style(TEST_FILE("bad_fonts/style_linking_issues/NotoSans-Bold.ttf")) == "Bold"
-    assert style(TEST_FILE("bad_fonts/style_linking_issues/NotoSans-BoldItalic.ttf")) == "BoldItalic"
-    # Badly named statics, fail them
-    assert style(TEST_FILE("bad_fonts/bad_stylenames/NotoSans-Fat.ttf")) == None
-    assert style(TEST_FILE("bad_fonts/bad_stylenames/NotoSans.ttf")) == None
 
 
 def test_check_name_copyright_length():
